Should work also with HTTPS base URL
If default.extensions.Drupal\MinkExtension.base_url is set to a URL using
HTTPS, $GLOBALS['base_url'] still returns http as scheme. On the following step,
an HTTPS request fails to get the cookie as it has been set previously with HTTP.
Use the behat.yml base URL to compute the reset link.